Use more descriptive names for active state tracking
