+++ lib/Drawer/Drawer.js
@@ -2,8 +2,8 @@ import {
   StateContainer,
   contentFactory,
   triggerFactory,
-  withActive,
-  withState
+  withActiveState,
+  withActive
 } from '../State'
 
 /**
@@ -16,10 +16,10 @@ import {
  * @constructor
  * @extends React.Component
  */
-export default withState({ type: 'drawer' })(
+export default withActive({ element: 'drawer' })(
   class Drawer extends StateContainer {
-    static Content = withActive({ id: true, hidden: true })(contentFactory())
-    static Toggle = withActive({ controls: true, expanded: true })(
+    static Content = withActiveState({ id: true, hidden: true })(contentFactory())
+    static Toggle = withActiveState({ controls: true, expanded: true })(
       triggerFactory({ link: true })
     )
   }
