+++ lib/Dropdown/Dropdown.js
@@ -2,8 +2,8 @@ import {
   StateContainer,
   contentFactory,
   triggerFactory,
-  withActive,
-  withState
+  withActiveState,
+  withActive
 } from '../State'
 import Button from '../Button'
 import elementFactory from '../utils/element-factory'
@@ -16,12 +16,12 @@ import elementFactory from '../utils/element-factory'
  * - Default focus on open first item
  * - Keydown listener for arrows to navigate through menu items
  */
-export default withState({ type: 'dropdown' })(
+export default withActive({ element: 'dropdown' })(
   class Dropdown extends StateContainer {
-    static Content = withActive({ labelledby: true, hidden: true })(
+    static Content = withActiveState({ labelledby: true, hidden: true })(
       contentFactory()
     )
-    static Toggle = withActive({ expanded: true, haspopup: true, id: true })(
+    static Toggle = withActiveState({ expanded: true, haspopup: true, id: true })(
       triggerFactory({ link: false })
     )
     static Item = elementFactory({
