+++ lib/Popover/Popover.js
@@ -2,8 +2,8 @@ import {
   StateContainer,
   contentFactory,
   triggerFactory,
-  withActive,
-  withState
+  withActiveState,
+  withActive
 } from '../State'
 
 /**
@@ -18,12 +18,12 @@ import {
  * @constructor
  * @extends React.Component
  */
-export default withState({ type: 'popover' })(
+export default withActive({ element: 'popover' })(
   class Popover extends StateContainer {
-    static Content = withActive({ id: true, role: 'tooltip', hidden: true })(
+    static Content = withActiveState({ id: true, role: 'tooltip', hidden: true })(
       contentFactory({ tip: true })
     )
-    static Toggle = withActive({ describedby: true, subscribe: false })(
+    static Toggle = withActiveState({ describedby: true, subscribe: false })(
       triggerFactory({ mouseEvents: true })
     )
   }
