+++ lib/State/index.js
@@ -2,21 +2,21 @@ import StateContainer from './StateContainer'
 import contentFactory from './content-factory'
 import triggerFactory from './trigger-factory'
 import withActive from './withActive'
-import withState from './withState'
+import withActiveState from './withActiveState'
 
-const State = withState()(StateContainer)
+const State = withActive()(StateContainer)
 
-State.Content = withActive({ id: true, hidden: true })(
-  contentFactory({ type: 'state' })
+State.Content = withActiveState({ id: true, hidden: true })(
+  contentFactory({ element: 'state' })
 )
 
-State.Toggle = withActive({ controls: true })(
+State.Toggle = withActiveState({ controls: true })(
   triggerFactory({ link: true, trigger: 'toggle' })
 )
-State.Activate = withActive({ controls: true })(
+State.Activate = withActiveState({ controls: true })(
   triggerFactory({ link: true, trigger: 'activate' })
 )
-State.Deactivate = withActive({ controls: true })(
+State.Deactivate = withActiveState({ controls: true })(
   triggerFactory({ link: true, trigger: 'deactivate' })
 )
 
@@ -26,5 +26,5 @@ export {
   contentFactory,
   triggerFactory,
   withActive,
-  withState
+  withActiveState
 }
