+++ lib/Tooltip/Tooltip.js
@@ -2,8 +2,8 @@ import {
   StateContainer,
   contentFactory,
   triggerFactory,
-  withActive,
-  withState
+  withActiveState,
+  withActive
 } from '../State'
 
 /**
@@ -15,12 +15,12 @@ import {
  * @constructor
  * @extends React.Component
  */
-export default withState({ type: 'tooltip' })(
+export default withActive({ element: 'tooltip' })(
   class Tooltip extends StateContainer {
-    static Content = withActive({ id: true, role: 'tooltip', hidden: true })(
+    static Content = withActiveState({ id: true, role: 'tooltip', hidden: true })(
       contentFactory({ tip: true })
     )
-    static Toggle = withActive({ describedby: true, subscribe: false })(
+    static Toggle = withActiveState({ describedby: true, subscribe: false })(
       triggerFactory({ mouseEvents: true })
     )
   }
